+++ apps/chrome-extension/src/pages/shared/settings.ts
@@ -23,7 +23,9 @@ export const useSettings = <
     | SettingKey.Theme
     | SettingKey.DownloadMethod
     | SettingKey.TableWithNonPhrasingContent
-    | SettingKey.TextHighlight,
+    | SettingKey.TextHighlight
+    | SettingKey.DownloadFileWithUniqueName
+    | SettingKey.FlatGrid,
 >(
   options: { keys?: Key[] } = {},
 ): {
@@ -42,6 +44,8 @@ export const useSettings = <
       SettingKey.DownloadMethod,
       SettingKey.TableWithNonPhrasingContent,
       SettingKey.TextHighlight,
+      SettingKey.DownloadFileWithUniqueName,
+      SettingKey.FlatGrid,
     ] as Key[],
   } = options