+++ apps/chrome-extension/src/scripts/view-lark-docx-as-markdown.ts
@@ -5,7 +5,10 @@ import { isDefined } from '@dolphin/common'
 import { CommonTranslationKey, en, Namespace, zh } from '../common/i18n'
 import { confirm } from '../common/notification'
 import { reportBug } from '../common/issue'
-import { transformInvalidTablesToHtml } from '../common/utils'
+import {
+  transformInvalidTablesToHtml,
+  transformMentionUsers,
+} from '../common/utils'
 import {
   getSettings,
   SettingKey,
@@ -83,11 +86,15 @@ const main = async () => {
   const settings = await getSettings([
     SettingKey.TableWithNonPhrasingContent,
     SettingKey.TextHighlight,
+    SettingKey.FlatGrid,
   ])
 
-  const { root, images, invalidTables } = docx.intoMarkdownAST({
+  const { root, images, invalidTables, mentionUsers } = docx.intoMarkdownAST({
     highlight: settings[SettingKey.TextHighlight],
+    flatGrid: settings[SettingKey.FlatGrid],
   })
+
+  await transformMentionUsers(mentionUsers)
 
   const tokens = images
     .map(image => {
@@ -107,7 +114,9 @@ const main = async () => {
     settings[SettingKey.TableWithNonPhrasingContent] ===
     TableWithNonPhrasingContent.ToHTML
   ) {
-    transformInvalidTablesToHtml(invalidTables)
+    transformInvalidTablesToHtml(invalidTables, {
+      allowDangerousHtml: true,
+    })
   }
 
   const markdown = Docx.stringify(root)